% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/directory_scaffold.R
\name{dirs.setup_shell}
\alias{dirs.setup_shell}
\title{Build a directory structure}
\usage{
dirs.setup_shell(set_scaffold = NULL)
}
\description{
\code{dirs.setup_shell}
}
\examples{
set_scaffold <- list(
 inst = list(
   www = list('style', 'images', 'lib', 'js'),
   "data")
 )

set_scaffold2 <- list(
inst = list(
  www = list(
    "style",
    "js",
    "images",
    "lib",
    "ref_tables",
    "templates",
    "shiny"
  ),

  data = list("locker")
),
revdep = list("scraps"))


}
