% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docr_author.R
\name{docr.author}
\alias{docr.author}
\title{Build an RD formatted Author line}
\usage{
docr.author(full_name = NULL, email = NULL)
}
\value{
The formatted roxygen line for the author.
}
\description{
\code{docr.author}
}
\details{
Set the author line for any documentation or R scripts.
}
\examples{
me <- list(name = "Carl S.A. Boneri", email = "carl.boneri@whyles.com")
docr.author(me$name, me$email)

# The function handles improperly specified email addresses containing only
# one \\code{'@'} symbol.

#' @author Carl S.A. Boneri, \\email{carl.boneri@whyles.com}


}
\author{
Carl S.A. Boneri, \email{carl.boneri@whyles.com}
}
\seealso{
Other Documentation.functions: \code{\link{docr.clean_fun}},
  \code{\link{docr.defaults_prompt}},
  \code{\link{docr.defaults}},
  \code{\link{docr.describe_df}},
  \code{\link{docr.example_df}}, \code{\link{docr.family}},
  \code{\link{docr.fun}}, \code{\link{docr.has_defaults}},
  \code{\link{docr.make_license}},
  \code{\link{docr.meta_private}},
  \code{\link{docr.meta_public}}, \code{\link{docr.param}}
}

