% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docr_defaults_prompt.R
\name{docr.defaults_prompt}
\alias{docr.defaults_prompt}
\title{Console prompting for docr default option settings}
\usage{
docr.defaults_prompt(...)
}
\description{
\code{docr.defaults_prompt}
}
\seealso{
Other Documentation.functions: \code{\link{docr.author}},
  \code{\link{docr.clean_fun}},
  \code{\link{docr.defaults}},
  \code{\link{docr.describe_df}},
  \code{\link{docr.example_df}}, \code{\link{docr.family}},
  \code{\link{docr.fun}}, \code{\link{docr.has_defaults}},
  \code{\link{docr.make_license}},
  \code{\link{docr.meta_private}},
  \code{\link{docr.meta_public}}, \code{\link{docr.param}}
}

