% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docr_describe_df.R
\name{docr.describe_df}
\alias{docr.describe_df}
\title{Build an RD formatted output of a dataframe object}
\usage{
docr.describe_df(df_in = NULL, make_interactive = FALSE,
  include_slice = FALSE)
}
\arguments{
\item{df_in}{The data frame object to be documented}

\item{make_interactive}{logical Should the columns of the data frame be
additionally documented. If set to \code{\link{TRUE}}, this will create a
prompted loop at the terminal, in which the column name will be printed, and
the provided information can be typed for the column information. The info
is then passed into the item description returned.}
}
\value{
A properly formatted roxygen block describing the data.frame in detail.
}
\description{
\code{docr.describe_df}
}
\details{
This is a helper function that works similarly to \code{\link{promptData}},
 but instead of returning an .RD formatted output file, this function
 \code{\link{cat}}'s a preformatted roxygenized block of information for use
 in documenting a data frame. The intention of this function is that when an
 alternate or newly created function returns a data.frame object, this function
 can be used to easily document the \strong{@return} code-block.
}
\examples{

# A data frame of all local RDS formatted files.

a <- dirs.get_r(find_pattern = ".rds$")

docr.describe_df(a, make_interactive = T)

#' \\format{
#'   A data frame with 19 observations on the following 5 variables.
#'  \\describe{
#'     \\item{\\strong{base_path}}{\\emph{a character vector}
#'         The home path to the library location.}
#'     \\item{\\strong{lib_name}}{\\emph{a character vector}
#'      The library name of the containing package or
#'      directory}
#'     \\item{\\strong{folder_path}}{\\emph{a character vector}
#'       The sub folder paths, within the directory, in
#'        which the file is stored}
#'     \\item{\\strong{file_name}}{\\emph{a character vector}
#'        The actual, basename of the file itself}
#'     \\item{\\strong{full_path}}{\\emph{a character vector}
#'         The fully normalized path to the file, for use
#'         when sourcing or reading in from an external or
#'         alternate library.}
#'  }
#'}

docr.describe_df(a, make_interactive = F)

#' \\format{
#'  A data frame with 19 observations on the following 5 variables.
#'  \\describe{
#'    \\item{\\strong{base_path}}{\\emph{a character vector}}
#'    \\item{\\strong{lib_name}}{\\emph{a character vector}}
#'    \\item{\\strong{folder_path}}{\\emph{a character vector}}
#'    \\item{\\strong{file_name}}{\\emph{a character vector}}
#'    \\item{\\strong{full_path}}{\\emph{a character vector}}
#'   }
#' }

}
\author{
Carl S.A. Boneri, \email{carl.boneri@whyles.com}
}
\seealso{
Other Documentation.functions: \code{\link{docr.author}},
  \code{\link{docr.clean_fun}},
  \code{\link{docr.defaults_prompt}},
  \code{\link{docr.defaults}},
  \code{\link{docr.example_df}}, \code{\link{docr.family}},
  \code{\link{docr.fun}}, \code{\link{docr.has_defaults}},
  \code{\link{docr.make_license}},
  \code{\link{docr.meta_private}},
  \code{\link{docr.meta_public}}, \code{\link{docr.param}}
}

