% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docr_example_df.R
\name{docr.example_df}
\alias{docr.example_df}
\title{Print a data frame for RD documentation in examples}
\usage{
docr.example_df(df, just_head = T, comments = NULL, write_fun = NULL)
}
\description{
\code{docr.example_df}
}
\examples{

 > docr.example_df(mtcars)

 #'=================  ====  ===  ====  ===  ====
 #'row_idx             mpg  cyl  disp   hp  drat
 #'=================  ====  ===  ====  ===  ====
 #'Mazda RX4          21.0    6   160  110  3.90
 #'Mazda RX4 Wag      21.0    6   160  110  3.90
 #'Datsun 710         22.8    4   108   93  3.85
 #'Hornet 4 Drive     21.4    6   258  110  3.08
 #'Hornet Sportabout  18.7    8   360  175  3.15
 #'Valiant            18.1    6   225  105  2.76
 #'=================  ====  ===  ====  ===  ====

 # NOW SET COMMENTS ON THE DATA FRAME ITSELF

> docr.example_df(mtcars, comments  = "these are cars")
# these are cars

=================  ====  ===  ====  ===  ====
row_idx             mpg  cyl  disp   hp  drat
=================  ====  ===  ====  ===  ====
Mazda RX4          21.0    6   160  110  3.90
Mazda RX4 Wag      21.0    6   160  110  3.90
Datsun 710         22.8    4   108   93  3.85
Hornet 4 Drive     21.4    6   258  110  3.08
Hornet Sportabout  18.7    8   360  175  3.15
Valiant            18.1    6   225  105  2.76
=================  ====  ===  ====  ===  ====
}
\seealso{
Other Documentation functions: \code{\link{docr.author}},
  \code{\link{docr.clean_fun}},
  \code{\link{docr.defaults_prompt}},
  \code{\link{docr.defaults}},
  \code{\link{docr.describe_df}},
  \code{\link{docr.family}}, \code{\link{docr.fun}},
  \code{\link{docr.has_defaults}},
  \code{\link{docr.install_r_only}},
  \code{\link{docr.make_license}},
  \code{\link{docr.meta_private}},
  \code{\link{docr.meta_public}}, \code{\link{docr.param}}
}
\author{
Carl S.A. Boneri, \email{carl.boneri@whyles.com}
}
