% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docr_family.R
\name{docr.family}
\alias{docr.family}
\title{Set a formatted family name}
\usage{
docr.family(family_name = NULL)
}
\arguments{
\item{family_name}{string The name of the family the function belongs to.}
}
\value{
A formatted roxygen string
}
\description{
\code{docr.family}
}
\examples{

}
\author{
Carl S.A. Boneri, \email{carl.boneri@whyles.com}
}
\seealso{
Other Documentation.functions: \code{\link{docr.author}},
  \code{\link{docr.clean_fun}},
  \code{\link{docr.defaults_prompt}},
  \code{\link{docr.defaults}},
  \code{\link{docr.describe_df}},
  \code{\link{docr.example_df}}, \code{\link{docr.fun}},
  \code{\link{docr.has_defaults}},
  \code{\link{docr.make_license}},
  \code{\link{docr.meta_private}},
  \code{\link{docr.meta_public}}, \code{\link{docr.param}}
}

