% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docr_freetype.R
\name{docr.freetype}
\alias{docr.freetype}
\title{Move all docr files into the working projects revdep folder}
\usage{
docr.freetype(file_name = NULL, assign_as = NULL)
}
\description{
Move all docr files into the working projects revdep folder
}

