% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docr_fun.R
\name{docr.fun}
\alias{docr.fun}
\title{Formetting for functions in alternate packages}
\usage{
docr.fun(...)
}
\description{
\code{docr.fun}
}
\examples{
> docr.fun(grep)
[1] "\\\\code{\\\\link[base]{grep}}"
> docr.fun(rt.css_prep_idx)
[1] "\\\\code{\\\\link[rtspecs]{rt.css_prep_idx}}"
> docr.fun(select)
[1] "\\\\code{\\\\link[dplyr]{select}}"


}
\author{
Carl S.A. Boneri, \email{carl.boneri@whyles.com}
}
\seealso{
Other Documentation.functions: \code{\link{docr.author}},
  \code{\link{docr.clean_fun}},
  \code{\link{docr.defaults_prompt}},
  \code{\link{docr.defaults}},
  \code{\link{docr.describe_df}},
  \code{\link{docr.example_df}}, \code{\link{docr.family}},
  \code{\link{docr.has_defaults}},
  \code{\link{docr.make_license}},
  \code{\link{docr.meta_private}},
  \code{\link{docr.meta_public}}, \code{\link{docr.param}}
}

