% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docr_get_defaults.R
\name{docr.get_defaults}
\alias{docr.get_defaults}
\title{Extract the global options associated with docr user}
\usage{
docr.get_defaults(field = NULL)
}
\value{
\format{
 A data frame with 1 observations on the following 5 variables.
 \describe{
   \item{\strong{author}}{\emph{character} formatted vector with \emph{1}
   unique values.}
   \item{\strong{email}}{\emph{character} formatted vector with \emph{1}
   unique values.}
   \item{\strong{company}}{\emph{character} formatted vector with \emph{1}
    unique values.}
   \item{\strong{website}}{\emph{character} formatted vector with \emph{1}
   unique values.}
   \item{\strong{github}}{\emph{character} formatted vector with \emph{1}
    unique values.}
  }
}


 @family Documentation functions
}
\description{
\code{docr.get_defaults}
}
