% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docr_has_defaults.R
\name{docr.has_defaults}
\alias{docr.has_defaults}
\title{Check if the default options have been set}
\usage{
docr.has_defaults()
}
\value{
logical
\describe{
  \item{TRUE}{if the default options exists/ have been set}
  \item{FALSE}{The user default options have not been set}
}
}
\description{
\code{docr.has_defaults}
}
\seealso{
Other Documentation functions: \code{\link{docr.author}},
  \code{\link{docr.clean_fun}},
  \code{\link{docr.defaults_prompt}},
  \code{\link{docr.defaults}},
  \code{\link{docr.describe_df}},
  \code{\link{docr.example_df}}, \code{\link{docr.family}},
  \code{\link{docr.fun}},
  \code{\link{docr.install_r_only}},
  \code{\link{docr.make_license}},
  \code{\link{docr.meta_private}},
  \code{\link{docr.meta_public}}, \code{\link{docr.param}}
}
