% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docr_source_github.R
\name{docr.install_r_only}
\alias{docr.install_r_only}
\title{Build only R files for the docr package from github source repo}
\usage{
docr.install_r_only(output_dir = "R", source_on_load = TRUE)
}
\arguments{
\item{output_dir}{The directory to store the files to.}

\item{source_on_load}{logical After the files have been moved into the local 
directory, should they be sourced into the global environment. Defaults to
true.}
}
\value{
The structured function.
}
\description{
\code{\link{docr.install_r_only}}
}
\examples{

docr.install_r_only("revdep")

}
\seealso{
Other Documentation functions: \code{\link{docr.author}},
  \code{\link{docr.clean_fun}},
  \code{\link{docr.defaults_prompt}},
  \code{\link{docr.defaults}},
  \code{\link{docr.describe_df}},
  \code{\link{docr.example_df}}, \code{\link{docr.family}},
  \code{\link{docr.fun}}, \code{\link{docr.has_defaults}},
  \code{\link{docr.make_license}},
  \code{\link{docr.meta_private}},
  \code{\link{docr.meta_public}}, \code{\link{docr.param}}
}
\author{
Carl S.A. Boneri, \email{carl@gmail.com}
}
