% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docr_list.R
\name{docr.list}
\alias{docr.list}
\title{Create an RD list from an R list}
\usage{
docr.list(type = c("ordered", "unordered", "named"), L)
}
\arguments{
\item{type}{Type of either:
\describe{
 \item{ordered}{A numbered list}
 \item{unordered(bulleted)}{A bullet list}
 \item{named}{A pair list of named objects}
}}

\item{L}{The list; if pre-constructed}
}
\description{
\code{docr.list}
}
\examples{
A <- list(
       Linux = 'server or machine type running in a UNIX/Linux environment',
       Windows = 'Running win-32 or win64 environment',
       `Mac OS` = 'Running MACOS system type'
     )

 # Since all list variables are named, the function defaults to `named`
 # list-type
 > docr.list(L = A)
##' \\\\describe{
##'   \\\\item{Linux}{server or machine type running in a UNIX/Linux environment}
##'   \\\\item{Windows}{Running win-32 or win64 environment}
##'   \\\\item{Mac OS}{Running MACOS system type}
##' }

> docr.list(L = A, type = "unordered")
##' \\\\itemize{
##'   \\\\item server or machine type running in a UNIX/Linux environment
##'   \\\\item Running win-32 or win64 environment
##'   \\\\item Running MACOS system type
##' }

> docr.list(L = A, type = "ordered")
##' \\enumerate{
##'   \\item server or machine type running in a UNIX/Linux environment
##'   \\item Running win-32 or win64 environment
##'   \\item Running MACOS system type
##' }

}
