% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docr_make_license.R
\name{docr.make_license}
\alias{docr.make_license}
\title{Generate a LICENSE formatted to Rd specs}
\usage{
docr.make_license(owner_name = NULL, owner_org = NULL, year = "current",
  output_to = c("file", "console"))
}
\arguments{
\item{owner_name}{string Default to NULL Owner of the License.}

\item{owner_org}{string Default to NULL If the owner attributes
work to an Organization.}

\item{year}{character Default to current year.}

\item{output_to}{call Default to file and console Where should the output
print?}
}
\value{
Output is either a LICENSE file object in the local or working
 directory or printed to the console
}
\description{
\code{docr.make_lic}
}
\examples{
doc.make_lic(owner_name = 'Carl Boneri',owner_org = "FAS Advisory",
output_to = 'file')
#> #'@param owner_name string Default to NULL
#> #'@param owner_org string Default to NULL
#> #'@param year character Default to current
#> #'@param output_to call Default to file and console



}
\seealso{
Other Documentation.functions: \code{\link{docr.author}},
  \code{\link{docr.clean_fun}},
  \code{\link{docr.defaults_prompt}},
  \code{\link{docr.defaults}},
  \code{\link{docr.describe_df}},
  \code{\link{docr.example_df}}, \code{\link{docr.family}},
  \code{\link{docr.fun}}, \code{\link{docr.has_defaults}},
  \code{\link{docr.meta_private}},
  \code{\link{docr.meta_public}}, \code{\link{docr.param}}
}

