% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docr_meta_private.R
\name{docr.meta_private}
\alias{docr.meta_private}
\title{This is a boilerplate meta template for working with private or patent projects.}
\usage{
docr.meta_private(user_name = NULL, user_email = NULL, org_name = NULL,
  org_website = NULL)
}
\description{
\code{docr.meta_private}
}
\details{
Any meta information, or function that by default returns a
 pre-formatted meta block of text to be included in the head
 or footer of a script file.
}
\seealso{
Other Documentation functions: \code{\link{docr.author}},
  \code{\link{docr.clean_fun}},
  \code{\link{docr.defaults_prompt}},
  \code{\link{docr.defaults}},
  \code{\link{docr.describe_df}},
  \code{\link{docr.example_df}}, \code{\link{docr.family}},
  \code{\link{docr.fun}}, \code{\link{docr.has_defaults}},
  \code{\link{docr.install_r_only}},
  \code{\link{docr.make_license}},
  \code{\link{docr.meta_public}}, \code{\link{docr.param}}
}
