% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docr_meta_public.R
\name{docr.meta_public}
\alias{docr.meta_public}
\title{Boilerplate header meta for scripts and source code for open-source packages}
\usage{
docr.meta_public(package_name = NULL, authors = NULL, years = NULL,
  to_file = NULL)
}
\description{
\code{docr.meta_public}
}
\seealso{
Other Documentation functions: \code{\link{docr.author}},
  \code{\link{docr.clean_fun}},
  \code{\link{docr.defaults_prompt}},
  \code{\link{docr.defaults}},
  \code{\link{docr.describe_df}},
  \code{\link{docr.example_df}}, \code{\link{docr.family}},
  \code{\link{docr.fun}}, \code{\link{docr.has_defaults}},
  \code{\link{docr.install_r_only}},
  \code{\link{docr.make_license}},
  \code{\link{docr.meta_private}}, \code{\link{docr.param}}
}
