% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docr_param.R
\name{docr.param}
\alias{docr.param}
\title{Generate preformatted param roxygen lines from a function name.}
\usage{
docr.param(fun_name = NULL)
}
\description{
\code{docr.param}
}
\seealso{
Other Documentation functions: \code{\link{docr.author}},
  \code{\link{docr.clean_fun}},
  \code{\link{docr.defaults_prompt}},
  \code{\link{docr.defaults}},
  \code{\link{docr.describe_df}},
  \code{\link{docr.example_df}}, \code{\link{docr.family}},
  \code{\link{docr.fun}}, \code{\link{docr.has_defaults}},
  \code{\link{docr.install_r_only}},
  \code{\link{docr.make_license}},
  \code{\link{docr.meta_private}},
  \code{\link{docr.meta_public}}
}
