% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docr_remote_source.R
\name{docr.remote_source}
\alias{docr.remote_source}
\title{DOCR source files from github repo}
\usage{
docr.remote_source(author = NULL, email = NULL, company = NULL,
  meta = NULL, website = NULL, github = NULL)
}
\description{
\code{docr.remote_source}
}
\examples{
docr.remote_source(author = "Carl Boneri", email = "carl.boneri@whyles.com",
                   company = "Whyles Inc", meta = NULL,
                   website = "https://whyles.com", github = "CarlBoneri")


}
