% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docr_txt.R
\name{docr.txt}
\alias{docr.txt}
\title{Text formatting helpers for RD documentation}
\usage{
docr.txt(type = c("italics", "strong", "code", "preformatted", "url", "href",
  "email"), ...)
}
\arguments{
\item{...}{The text body}

\item{format}{The formatting variable for the text object; one of:

 \describe{
   \item{italics}{emphasises text in italics}
   \item{strong}{creates bold text}
   \item{code}{formats inline code}
   \item{preformatted}{format ext as-is, useful for multi-line code}
   \item{url}{a url or web reference}
   \item{href}{url with custom link}
   \item{email}{an email address}
 }}
}
\description{
\code{docr.txt}
}
\examples{
# Created a line of body text with formatting
my_text <- paste0(collapse=" ",
             c('A function which allows passing',
               docr.txt('italics','external'),
               'arguments in the form of',
               docr.txt('code','alist')))

> docr.wrap(my_text)
# #'  A function which allows passing \\\\emph{external} arguments
# #'   in the form of \\\\code{alist}


}

