% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{env.clear_console}
\alias{env.clear_console}
\title{Clean all lines out of the console.}
\usage{
env.clear_console()
}
\description{
\code{env.clear_console}
}
\details{
Rstudio Server can get really slow if a lot of text, especially unicode
 or foreign data types are sitting in the GUI. This function is the equiv
 of \code{Edit > Ctrl + L}
}
\seealso{
Other Environment utilities: \code{\link{env.clear_objects}}
}
