% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{env.clear_objects}
\alias{env.clear_objects}
\title{Function for clearing all objects from cache in a session}
\usage{
env.clear_objects(keep_pat = NULL, drop_pat = NULL)
}
\description{
\code{env.clear_objects}
}
\details{
Working in an environment or in an app can mean the workspace or
in-memory storage can get bogged down with unecessary data objects. This
can impact speed or simply clutter the space itself. This function will
remove all objects that are not functions in the current environment.
}
\examples{

# From a working environment
> sapply(ls(), function(i){
 is.function(get(i))
}) \%>\% {
  dput(names(.[!.]))
}

[1] "nf" "nf.raw" "nf.team_df" "nf.team_df_master" "nf.teams"

# Now clear anything beginning with nf followed by a dot and has a word after
# such as nf.team_df

> env.clear_objects(drop_pat = "nf\\\\.(\\\\w+)+$")

> sapply(ls(), function(i){
     is.function(get(i))
 }) \%>\% {
      (names(.[!.]))
 }

 [1] "nf"

}
\seealso{
Other Environment utilities: \code{\link{env.clear_console}}
}
