% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{env.idx}
\alias{env.idx}
\title{Function to index every object in an environment or namespace}
\usage{
env.idx(filter_class = NULL, env = "global", pattern = NULL)
}
\arguments{
\item{filter_class}{string or character vector, Pass the class of objects
to be returned, if NULL, the default, return all.}

\item{env}{string This variable interprets the input as a package name, or
the default, global, the working invironment. This is the environment from
which we want the objects from.}

\item{pattern}{A regular expression or string to find in the environment.

 @return
\format{
 A data frame with 77 observations on the following 2 variables.
 \describe{
   \item{\strong{obj_name}}{\emph{a character vector}
        This is the name of the object within the
        active, or called namespace.}
   \item{\strong{obj_type}}{\emph{a character vector}
        The class of the object. Such as character,
        data.frame, environment etc.}
  }
}}
}
\description{
\code{env.idx}
}
\details{
This function is helpful when space becomes cluttered, or even for finding
and calling alternate variables or functions from within a space
}
\examples{
============  ========
obj_name      obj_type
============  ========
.             function
aaply         function
adply         function
alply         function
amv_dimnames  function
a_ply         function
============  ========

}
\author{
Carl S.A. Boneri, \email{carl@gmail.com}
}
