% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/env_utils.R, R/utils.R
\name{env.new}
\alias{env.new}
\alias{env.new}
\title{Function to create new environments}
\usage{
env.new(env_name = NULL)

env.new(env_name = NULL)
}
\arguments{
\item{env_name}{String, the name of the new environment.}
}
\description{
Function to create new environments

\code{env.new}
}
\details{
Works much like calling \code{\link{new.env()}} but allows passing a string
argument to assign to the environment namespace. Useful within function calls
where a new env can be used to mask operations and send outputs early to
outside functions while the intended function completes its steps.
}
\examples{

env.new('.hidden')

}
