% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{files.download}
\alias{files.download}
\title{Download a file from the internet and copy to a local directory}
\usage{
files.download(src_path = NULL, dest_name = NULL, dest_path = NULL,
  multi = FALSE, in_background = TRUE)
}
\description{
\code{files.download}
}
