% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{files.unzip}
\alias{files.unzip}
\title{Download and unzip a file from the internet or other external resource}
\usage{
files.unzip(zip_path = NULL, dest_path = NULL, keep_src = FALSE, ...)
}
\arguments{
\item{zip_location}{The url, or directory path pointing to the
location of the source zip file.}

\item{exdir}{The directory to extract the the contents of the zip folder to.}

\item{pattern}{If set, this is a regular expression pattern to be used
for telling the function which files to keep. The side-effect is that all
files not matching will be discarded.}
}
\description{
\code{files.unzip}
}
