% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{pound}
\alias{pound}
\title{Create a string of #}
\usage{
pound(n = NULL, use_symbol = "#")
}
\description{
\code{pound}
}
\examples{

> pound(n = 14)
[1] "##############"
> pound(n = 14, "--")
[1] "----------------------------"
}
