% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docr_clean_fun.R
\name{docr.clean_fun}
\alias{docr.clean_fun}
\title{Build a clean and set-formatted function}
\usage{
docr.clean_fun(fun_name = NULL, trim_blanks = FALSE, rm_comments = FALSE,
  do_cat = TRUE)
}
\arguments{
\item{fun_name}{The name of the function to be rebuilt}

\item{trim_blanks}{Should any carriage returns or empty lines be removed from
the source code.}

\item{rm_comments}{Should the comments in the source script of the function
be removed.}

\item{do_cat}{Should the output be printed to the console by calling
\code{\link{cat}}.}
}
\value{
The structured function.
}
\description{
\code{\link{docr.clean_fun}}
}
\details{
If a function is written at the command line, or in a file with no
documentation, this function will take the function name, in a quoted
variable state, and rebuild the function in the same-formatted structure
as when it was initially written.
}
\examples{

# From the Base package \\code{\\link{grep}}

> docr.clean_fun('grep')
grep <- function (pattern, x, ignore.case = FALSE, perl = FALSE, value = FALSE,
                  fixed = FALSE, useBytes = FALSE, invert = FALSE)
{
  if (!is.character(x))
    x <- structure(as.character(x), names = names(x))
  .Internal(grep(as.character(pattern), x, ignore.case, value,
                 perl, fixed, useBytes, invert))
}

# From the command line..we create the following:
paste.html <- function(x){
   htmltools::HTML(paste0(x, collapse = "\\n"))
}

# Notice the output is the same as the written style, to match the author's
# preference.

> docr.clean_fun('paste.html')
paste.html <- function(x){
 htmltools::HTML(paste0(x, collapse = "\\n"))
}

}
\seealso{
Other Documentation functions: \code{\link{docr.author}},
  \code{\link{docr.defaults_prompt}},
  \code{\link{docr.defaults}},
  \code{\link{docr.describe_df}},
  \code{\link{docr.example_df}}, \code{\link{docr.family}},
  \code{\link{docr.fun}}, \code{\link{docr.has_defaults}},
  \code{\link{docr.install_r_only}},
  \code{\link{docr.make_license}},
  \code{\link{docr.meta_private}},
  \code{\link{docr.meta_public}}, \code{\link{docr.param}}
}
\author{
Carl S.A. Boneri, \email{carl@gmail.com}
}
